requirements.txt
a2wsgi==1.10.10
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==5.1.0
aiosqlite==0.21.0
alembic==1.13.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==3.1.7
apache-airflow-core==3.1.7
apache-airflow-providers-apache-spark==5.5.1
apache-airflow-providers-common-compat==1.14.0
apache-airflow-providers-common-io==1.7.1
apache-airflow-providers-common-sql==1.32.0
apache-airflow-providers-fab==1.2.2
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==2.4.2
apache-airflow-providers-sqlite==3.8.1
apache-airflow-providers-standard==1.12.0
apache-airflow-task-sdk==1.1.7
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==25.4.0
azure-core==1.39.0
azure-storage-blob==12.28.0
babel==2.18.0
blinker==1.8.2
cachelib==0.9.0
cadwyn==5.4.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.3.1
clickclick==20.10.2
colorama==0.4.6
colorlog==6.10.1
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.3
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
duckdb==1.4.4
email_validator==2.2.0
Faker==40.12.0
fastapi==0.117.1
fastapi-cli==0.0.24
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.61.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240702
googleapis-common-protos==1.72.0
greenback==1.3.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.7.1
httpx==0.27.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kafka-python==2.3.0
kiwisolver==1.4.9
lazy-object-proxy==1.10.0
libcst==1.8.6
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.8
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
msgspec==0.20.0
multidict==6.0.5
natsort==8.4.0
numpy==2.4.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
pathspec==0.12.1
pendulum==3.2.0
pillow==12.1.1
pluggy==1.5.0
prison==0.2.1
protobuf==6.33.5
psutil==7.2.2
py4j==0.10.9.9
pyarrow==23.0.1
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
pygtrie==2.5.0
PyJWT==2.11.0
pyparsing==3.3.2
pyspark==4.1.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rich-toolkit==0.19.7
rpds-py==0.19.0
scikit-learn==1.8.0
scipy==1.17.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.5
starlette==0.48.0
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==8.5.0
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.14.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.0.7
uuid6==2025.0.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
